Do not use --proxy-user. Use HADOOP_USER_NAME.
In theory, both should be idential. Calling

    org.apache.hadoop.security.UserGroupInformation.getCurrentUser()

Produces almost identical results, e.g.:

    vladimir (auth:PROXY) via zeppelin (auth:SIMPLE)
    vladimir (auth:SIMPLE)

However, with the first user connection to HBase simply does not work. The
symptom is that it fails with ConnectionClosingException, configuring
proxy user on hbase does not appear to work, and there is a bug

  https://issues.apache.org/jira/browse/HBASE-12230

that might be related. So, just use another way to specify the user.