Clean up websocket on error.
Jetty websocket can emit either onClose or onError event. Previously,
the onClose event results in proper clean-up of websocket, while onError
was ignored - the socket would remain in the list of connections,
and all broadcasts would try to send data again to it, resulting in
yet another error that would be ignored.