Reassociate note with connection after reconnecting.
Turns out that Websocket protocol in Zeppelin is stateful, and
calls such as "run paragraph" don't take note id explicitly, but rather
assume that connection is already associated with note id, e.g. via
prior "get note" call.

When we remove the full-reload-on-reconnect behaviour, we also
removed the "get note" call and therefore the association is never
established. This commit adds explicit call to reassociate.